+++ protographic/CHANGELOG.md
@@ -4,6 +4,22 @@ Binaries are attached to the github release otherwise all images can be found [h
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.12.6](https://github.com/wundergraph/cosmo/compare/@wundergraph/protographic@0.12.5...@wundergraph/protographic@0.12.6) (2025-12-04)
+
+**Note:** Version bump only for package @wundergraph/protographic
+
+## [0.12.5](https://github.com/wundergraph/cosmo/compare/@wundergraph/protographic@0.12.4...@wundergraph/protographic@0.12.5) (2025-12-03)
+
+**Note:** Version bump only for package @wundergraph/protographic
+
+## [0.12.4](https://github.com/wundergraph/cosmo/compare/@wundergraph/protographic@0.12.3...@wundergraph/protographic@0.12.4) (2025-12-03)
+
+**Note:** Version bump only for package @wundergraph/protographic
+
+## [0.12.3](https://github.com/wundergraph/cosmo/compare/@wundergraph/protographic@0.12.2...@wundergraph/protographic@0.12.3) (2025-11-29)
+
+**Note:** Version bump only for package @wundergraph/protographic
+
 ## [0.12.2](https://github.com/wundergraph/cosmo/compare/@wundergraph/protographic@0.12.1...@wundergraph/protographic@0.12.2) (2025-11-26)
 
 **Note:** Version bump only for package @wundergraph/protographic